idea-hn-47353558-show-hn--onecli---vault-for-ai-agents-in/requirements.txt
Flask==3.0.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
